Move Overwritable to its own file, add todos
